Strip leading license header from source content
This is horrible. But AFAIK, we always want to do this.
And if we don't, we can add a flag later to toggle the behavior.
I just hope the behavior doesn't need to be customized more later.

I debated instead adding an option to give a line offset, but that
would not be robust against changes to the license headers.  E.g.,
if we switch from BSD-2 to ISC (basically a
shorter/simplified/nicer version of BSD-2), it would break
hardcoded offsets. Furthermore, with this hacky/clever regex, uses
of #github on the wiki do not need to manually specify an offset.